more structured code

+++ src/server/client.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <ctime>
 #include <string>
 #include <sys/socket.h>
@@ -6,9 +7,9 @@
 
 struct Client {
   int fd_main;
-  size_t client_id = 0;
-  size_t player_id = 0;
-  size_t room_id = 0;
+  uint32_t client_id = 0;
+  uint32_t player_id = 0;
+  uint32_t room_id = 0;
   std::time_t last_response;
   int epd = -1;
   int todo_fd = -1;
